+++ EDDI/ChangeLog.md
@@ -3,6 +3,7 @@
 ### 3.0.0-rc2
   * Core
     * The EDSM responder has been updated to send data to EDSM per their revised API. 
+    * Switched error reporting to [Rollbar](https://rollbar.com/).
   * Speech Responder
     * Added 'Fighter rebuilt' event